+++ components/src/lib/index.ts
@@ -1,9 +1,11 @@
 import Title from './components/common/Title/Title';
 import { ModalBody, ModalFooter, ModalHeader, ModalMain } from './components/Modal/Modal';
 import Input from './components/common/Input/Input';
-import { AlertModal, ConfirmModal, PromptModal } from './ModalPreset/ModalPreset';
 import CloseButton from './components/common/CloseButton/CloseButton';
 import Button from './components/common/Button/Button';
+import { AlertModal } from './components/Modal/ModalPreset/AlertModal';
+import { ConfirmModal } from './components/Modal/ModalPreset/ConfirmModal';
+import { PromptModal } from './components/Modal/ModalPreset/PromptModal';
 
 export const Modal = Object.assign(ModalMain, {
   Title: Title,
